Create ProfileChiSquared algorithm and clean up CalculateChiSquared
This commit creates a ProfileChiSquaredAlgorithm. It builds upon the implementation within CalculateChiSquared, but corrects for a few approximations. This has been validated against LMFIT and analytical hand-calculations.
